Added news to nav_menu and fixed bugs
