+++ app/src/main/java/com/prometheo/moneylife/ui/news/NewsViewModel.kt
@@ -8,6 +8,7 @@ import com.prometheo.moneylife.data.models.TurnEvent
 import com.prometheo.moneylife.data.models.UserIdBody
 import com.prometheo.moneylife.data.preferences.Prefs
 import com.prometheo.moneylife.data.room.AppDatabase
+import com.prometheo.moneylife.data.room.TurnEventDao
 import com.prometheo.moneylife.data.services.TurnService
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
@@ -17,30 +18,11 @@ import javax.inject.Inject
 class NewsViewModel @Inject constructor (
     private val turnService: TurnService,
     private val prefs: Prefs,
-    private val db: AppDatabase
+    private val turnEventDao: TurnEventDao
 ) : ViewModel() {
 
-    val turnEvents: LiveData<List<TurnEvent>> = db.turnEventDao().observeAll()
+    val turnEvents: LiveData<List<TurnEvent>> = turnEventDao.observeAll()
 
     private val _loading = MutableLiveData<Boolean>()
     val loading: LiveData<Boolean> = _loading
-
-    //TODO: Mover a TurnViewModel?
-    fun updateTurnEvents () {
-
-        _loading.value = true
-
-        viewModelScope.launch {
-            try {
-                val response = turnService.getTurnEvents( UserIdBody( prefs.userId ) )
-                if ( response.isSuccessful ) {
-                    db.turnEventDao().insert( response.body()?.first()!! )
-                }
-            } catch ( err: Throwable ) {
-                //TODO: Add error message
-            }
-            _loading.value = false
-        }
-    }
-
 }
