+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -12,6 +12,11 @@
         android:title="@string/title_health" />
 
     <item
+        android:id="@+id/navigation_news"
+        android:icon="@drawable/ic_dashboard_black_24dp"
+        android:title="@string/title_news" />
+
+    <item
         android:id="@+id/navigation_dashboard"
         android:icon="@drawable/ic_dashboard_black_24dp"
         android:title="@string/title_dashboard" />
