+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -18,6 +18,12 @@
         tools:layout="@layout/fragment_health" />
 
     <fragment
+        android:id="@+id/navigation_news"
+        android:name="com.prometheo.moneylife.ui.news.NewsFragment"
+        android:label="@string/title_news"
+        tools:layout="@layout/fragment_dashboard" />
+
+    <fragment
         android:id="@+id/navigation_dashboard"
         android:name="com.prometheo.moneylife.ui.dashboard.DashboardFragment"
         android:label="@string/title_dashboard"
